temp zip/tc001-main/tc001-main/timeceptor01/requirements.txt
anyio==4.9.0
bcrypt==4.1.3
click==8.2.1
fastapi==0.115.12
fastapi-jwt-auth==0.5.0
greenlet==3.2.3
h11==0.16.0
idna==3.10
passlib==1.7.4
psycopg2-binary==2.9.10
pydantic==1.10.22
PyJWT==1.7.1
pyswisseph==2.10.3.2
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing_extensions==4.14.0
uvicorn==0.34.3
